+++ docs/README.md
@@ -1,24 +1,25 @@
 # 로직 흐름
 
-1. controller : main 메서드와 직접적으로 통신하며 전체 로직 흐름을 담당하는 controller
-    1. 어플리케이션 시작 메서드안에 반복문을 만들어 안에 게임 시작하는 메서드 호출한다
-    2. "게임 시작 메서드"에서 정답 값을 만들고, 정답이 맞을때까지 반복하는 반복문에 들어간다
-    3. 반복문에서 입력값을 받고 받은 입력값이 타당한지 판단한다
-    4. 타당하면 입력값이 정답인지 판단하고 아니면 계속 반복문을 돌고 아니면 "게임 시작 메서드를 나간다"
+1. controller : view, domain 흐름을 관리하는 부분
+    1. 게임 로직 총괄 컨트롤러
+        1. 어플리케이션 시작 메서드안에 반복문을 만들어 안에 게임 시작하는 메서드 호출한다
+        2. "게임 시작 메서드"에서 정답 값을 만들고, 정답이 맞을때까지 반복하는 반복문에 들어간다
+        3. 반복문에서 입력값을 받고 받은 입력값이 타당한지 판단한다
+        4. 타당하면 입력값이 정답인지 판단하고 아니면 계속 반복문을 돌고 아니면 "게임 시작 메서드를 나간다"
+    2. 게임 시작을 담당하는 컨트롤러 -> GameStartController
+        - 정답 숫자 생성하는 로직
+        - inputView에서 게임 시작 출력 메서드 호출
+    3. 게임 종료 관련 로직 담당 컨트롤러 -> GameEndController
+        - userball 객체 내부 값에 따라 정답인지 판단
+        - 정답이면, ouputview, inputview의 메서드 후출하고 true 를 반환
 2. domain : 핵심 도메인 model 부분
     - 스트라이크 개수, 볼 개수 필드
     - 볼개수 업데이트 하는 메서드
     - 스트라이크 개수 업데이트하는 메서드
-3. service : 순수한 비지니스 로직이 담겨 있는 곳
+3. service : 순수 도메인만 관리하는 부분
     1. 입력값을 판단한하는 역할 (핵심 로직) (스트라이크, 볼, 낫싱 등등)  -> UserBallService
         - 우선 유저 입력값에 따른 UserBall 생성
             - 이때 입력값에 따라 ball , strike가 몇개인지 판다하며 userball 업데이트 한다
-    2. 게임 시작을 담당하는 서비스 -> GameStartService
-        - 정답 숫자 생성하는 로직 -> userballservice
-        - inputView에서 게임 시작 출력 메서드 호출
-    3. 게임 종료 관련 로직 담당 서비스 -> GameEndService
-        - userball 객체 내부 값에 따라 정답인지 판단
-        - 정답이면, ouputview, inputview의 메서드 후출하고 true 를 반환
         - 정답이 아니면, 볼, 스타라이크 개수에 따라 Ouputview 출력문 호출하고 false 반환
 4. view : 사용자에게 화면을 보여주는 부분에 초점
     1. 입력을 받는 역할 (inputView)
@@ -39,32 +40,33 @@
 
 # 구체적인 기능(메서드)
 
-1. controller : main 메서드와 직접적으로 통신하며 전체 로직 흐름을 담당하는 controller
-    1. 어플리케이션 시작 메서드
-        - 반복문 안에 게임 시작하는 메서드 호출
-        - 종료 조건에 어플리케이션 종표 판단 메서드 호출
-    2. 게임이 시작하는 메서드 (여기 정답이 틀리면, 계속 돌아야 하는 반복문임 만약 정)
-        - 정답 숫자를 생성 -> service 한테 로직을 맡김
-        - 반복문
-            - 올바른 입력값을 판단하는 메서드 -> util 한테 로직을 맡김
-            - 정답인지 판단하는 메서드(= 종료 조건 판단) -> serivce 한테 로직을 맡김
-        - inputview의 의존성을 받아서 입력하기 전의 게임 설명 부분을 출력함
-    3. 어플리케이션 종료 판단 메서드
-        1. 사용자가 게임을 다시 할것이라고 하면 게임시작 true를 반환
-        2. 사용자가 게임을 종료 한다고 하면, 게임 종료 false 반환
-2. domain : 핵심 도메인 model 부분
+1. controller : view, domain 흐름을 관리하는 부분
+    1. 게임 로직 총괄 컨트롤러
+        1. 어플리케이션 시작 메서드
+            - 반복문 안에 게임 시작하는 메서드 호출
+            - 종료 조건에 어플리케이션 종표 판단 메서드 호출
+        2. 게임이 시작하는 메서드 (여기 정답이 틀리면, 계속 돌아야 하는 반복문임 만약 정)
+            - 정답 숫자를 생성 -> service 한테 로직을 맡김
+            - 반복문
+                - 올바른 입력값을 판단하는 메서드 -> util 한테 로직을 맡김
+                - 정답인지 판단하는 메서드(= 종료 조건 판단) -> serivce 한테 로직을 맡김
+            - inputview의 의존성을 받아서 입력하기 전의 게임 설명 부분을 출력함
+        3. 어플리케이션 종료 판단 메서드
+            1. 사용자가 게임을 다시 할것이라고 하면 게임시작 true를 반환
+            2. 사용자가 게임을 종료 한다고 하면, 게임 종료 false 반환
+    2. 게임 시작을 담당하는 컨트롤러 -> GameStartController
+        - 정답 숫자 생성하는 로직
+        - inputView에서 게임 시작 출력 메서드 호출
+    3. 게임 종료 관련 로직 담당 컨트롤러 -> GameEndController
+        - userball에 따른 OutView 호출하는 메서드
+4. domain : 핵심 도메인 model 부분
     - 스트라이크 개수, 볼 개수 필드
     - 볼개수 업데이트 하는 메서드
     - 스트라이크 개수 업데이트하는 메서드
-4. service : 순수한 비지니스 로직이 담겨 있는 곳
+5. service : 순수 도메인만 관리하는 부분
     1. 입력값을 판단한하는 역할 (핵심 로직) (스트라이크, 볼, 낫싱 등등)  -> UserBallService
         - UserBall 객체를 생성 및 입력값에 따른 ball, strike 업데이트 메서드
-    2. 게임 시작을 담당하는 서비스 -> GameStartService
-        - 정답 숫자 생성하는 로직 -> userballservice
-        - inputView에서 게임 시작 출력 메서드 호출
-    3. 게임 종료 관련 로직 담당 서비스 -> GameEndService
-        - userball에 따른 OutView 호출하는 메서드
-5. view : 사용자에게 화면을 보여주는 부분에 초점
+6. view : 사용자에게 화면을 보여주는 부분에 초점
     1. 입력을 받는 역할 (inputView)
         - 게임 시작 문구 출력하는 메서드
         - 정답 값을 맞추고 게임을 종료할 것인지 사용자한테 출력문 출력 메서드
@@ -74,11 +76,11 @@
         - 스트라이크를 출력하는 메서드
         - 정답을 맞추면 출력하는 메서드
         - 낫싱을 출력하는 메서드
-6. util : 유용한 패키지 (외부의 의존성 없이 인자값만 받아 처리하는 부분, 에러 상수값)
+7. util : 유용한 패키지 (외부의 의존성 없이 인자값만 받아 처리하는 부분, 에러 상수값)
     - 예외를 처리하는 상수 -> errorConst
     - inputView, outputView에서 사용하는 상수 -> viewConst
     - 입력값을 이 valid 한지 판단하는 로직 -> inputvalidUtil
-7. config :
+8. config :
     - 사용할 객체 생성 및 생성자 주입
