+++ src/main/java/baseball/Application.java
@@ -1,12 +1,12 @@
 package baseball;
 
 import baseball.config.GameConfig;
-import baseball.controller.FrontController;
+import baseball.controller.GameCenterController;
 
 public class Application {
 	public static void main(String[] args) {
 		// TODO: 프로그램 구현
-		FrontController frontController = new GameConfig().config();
-		frontController.startApplication();
+		GameCenterController gameCenterController = new GameConfig().config();
+		gameCenterController.startApplication();
 	}
 }
