+++ src/main/java/baseball/config/GameConfig.java
@@ -1,8 +1,8 @@
 package baseball.config;
 
-import baseball.controller.FrontController;
-import baseball.service.GameEndService;
-import baseball.service.GameStartService;
+import baseball.controller.GameCenterController;
+import baseball.controller.GameEndController;
+import baseball.controller.GameStartController;
 import baseball.service.UserBallService;
 import baseball.util.InputValidUtil;
 import baseball.view.InputView;
@@ -10,17 +10,17 @@ import baseball.view.OutputView;
 
 public class GameConfig {
 
-	public FrontController config() {
+	public GameCenterController config() {
 
 		InputView inputView = new InputView();
 		OutputView outputView = new OutputView();
 
-		GameStartService gameStartService = new GameStartService(inputView);
+		GameStartController gameStartController = new GameStartController(inputView);
 		InputValidUtil inputValidUtil = new InputValidUtil();
 		UserBallService userBallService = new UserBallService();
-		GameEndService gameEndService = new GameEndService(outputView, inputView);
+		GameEndController gameEndController = new GameEndController(outputView, inputView);
 
-		return new FrontController(gameStartService, userBallService, gameEndService, inputValidUtil);
+		return new GameCenterController(gameStartController, userBallService, gameEndController, inputValidUtil);
 	}
 
 }
